+++ src/main/java/com/revolvingmadness/sculk/Sculk.java
@@ -1,21 +1,28 @@
 package com.revolvingmadness.sculk;
 
 import com.revolvingmadness.sculk.gamerules.SculkGamerules;
+import com.revolvingmadness.sculk.keybinds.KeybindHelper;
+import com.revolvingmadness.sculk.language.Event;
 import com.revolvingmadness.sculk.language.EventHolder;
+import com.revolvingmadness.sculk.language.builtins.classes.instances.entity.ServerPlayerEntityInstance;
 import com.revolvingmadness.sculk.language.lexer.TokenType;
 import net.fabricmc.api.ModInitializer;
 import net.fabricmc.fabric.api.event.lifecycle.v1.ServerLifecycleEvents;
+import net.fabricmc.fabric.api.networking.v1.ServerPlayNetworking;
+import net.minecraft.client.util.InputUtil;
 import net.minecraft.server.MinecraftServer;
 import net.minecraft.util.Identifier;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
 
 import java.util.HashMap;
+import java.util.List;
 import java.util.Map;
 
 public class Sculk implements ModInitializer {
     public static final String ID = "sculk";
     public static final Identifier DYNAMIC_REGISTRY_SYNC_ID = new Identifier(Sculk.ID, "dynamic_registry_sync");
+    public static final Identifier KEY_PRESS_ID = new Identifier(Sculk.ID, "key_press");
     public static final Logger LOGGER = LoggerFactory.getLogger(Sculk.ID);
     public static final Map<String, TokenType> keywords = new HashMap<>();
     public static MinecraftServer server;
@@ -25,7 +32,7 @@ public class Sculk implements ModInitializer {
             return 255;
         }
 
-        return Sculk.getMaxArgumentCount();
+        return Sculk.server.getGameRules().getInt(SculkGamerules.MAX_ARGUMENT_COUNT);
     }
 
     public static int getMaxLoops() {
@@ -42,6 +49,20 @@ public class Sculk implements ModInitializer {
         EventHolder.registerEvents();
 
         ServerLifecycleEvents.SERVER_STARTING.register(server -> Sculk.server = server);
+
+        ServerPlayNetworking.registerGlobalReceiver(Sculk.KEY_PRESS_ID, (server, player, handler, buf, responseSender) -> {
+            InputUtil.Key key = PacketByteBufSerialization.readKey(buf);
+
+            if (key.getCategory() != InputUtil.Type.KEYSYM) {
+                return;
+            }
+
+            List<Event> events = KeybindHelper.EVENTS.getOrDefault(key.getCode(), List.of());
+
+            for (Event event : events) {
+                event.execute(List.of(new ServerPlayerEntityInstance(player)));
+            }
+        });
 
         // Values
         Sculk.keywords.put("true", TokenType.TRUE);